LV: Export parts of pipeline
